requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
eval_type_backport==0.2.2
filelock==3.17.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
joblib==1.4.2
lxml==5.3.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
numpy==2.2.3
openai==1.65.2
pillow==11.1.0
propcache==0.3.0
psycopg2-binary==2.9.10
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
tabulate==0.9.0
threadpoolctl==3.5.0
tiktoken==0.9.0
together==1.4.1
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
